services/strategy/pyproject.toml
[project]
name = "strategy-service"
version = "0.1.0"
description = "BLOASIS Strategy Service - Stock Selection Pipeline Stage 3"
requires-python = ">=3.11"

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["/workspace"]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
]

[tool.coverage.run]
omit = [
    "src/main.py",  # Server entry point, tested via E2E
    "src/clients/*",  # gRPC clients, tested via integration tests with real services
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".pytest_cache",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["shared"]
